requirements.txt
Django==3.2.21
djangorestframework==3.14.0
numpy==1.19.5
pandas==1.1.5
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-django==4.5.2
pytest==7.0.1
requests==2.27.1
